prevents login and register pages being accessible for logged in user - must sign out first. extracts logic of both  forms out to components in order to -use client- there
