+++ app/api/auth/[...nextauth]/route.ts
@@ -9,7 +9,7 @@ interface Credentials {
     password: string;
 }
 
-const authOptions: NextAuthOptions = {
+export const authOptions: NextAuthOptions = {
     providers: [
         CredentialsProvider({
             name: "credentials",
