+++ app/components/RegisterForm.tsx
@@ -1,0 +1,105 @@
+"use client";
+import Link from "next/link";
+import { useRouter } from "next/navigation";
+import React, { useState } from "react";
+
+const RegisterForm = () => {
+    const [name, setName] = useState("");
+    const [username, setUsername] = useState("");
+    const [email, setEmail] = useState("");
+    const [password, setPassword] = useState("");
+    const [error, setError] = useState("");
+
+    const router = useRouter();
+
+    const handleSubmit = async (e) => {
+        e.preventDefault();
+
+        if (!name || !username || !email || !password) {
+            setError("All fields are required.");
+            return;
+        }
+
+        try {
+            const resUserExists = await fetch("api/userExists", {
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json",
+                },
+                body: JSON.stringify({ email }),
+            });
+
+            const { user } = await resUserExists.json();
+
+            if (user) {
+                setError("User already exists.");
+                return;
+            }
+            const res = await fetch("api/register", {
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json",
+                },
+                body: JSON.stringify({
+                    name,
+                    username,
+                    email,
+                    password,
+                }),
+            });
+            if (res.ok) {
+                const form = e.target;
+                form.reset();
+                router.push("/login");
+            } else {
+                console.log("User registration failed");
+            }
+        } catch (error) {
+            console.log("Error during registration");
+        }
+    };
+
+    return (
+        <div className="grid place-items-center h-screen">
+            <div className="shadow-lg p-5 rounded-lgshadow-lg p-5 rounded-lg border-t-4 border-green-400">
+                <h1 className="text-xl font-bold my-4">Register an account:</h1>
+                <form onSubmit={handleSubmit} className="flex flex-col gap-3">
+                    <input
+                        onChange={(e) => setName(e.target.value)}
+                        type="text"
+                        placeholder="Full Name"
+                    />
+                    <input
+                        onChange={(e) => setUsername(e.target.value)}
+                        type="text"
+                        placeholder="Username"
+                    />
+                    <input
+                        onChange={(e) => setEmail(e.target.value)}
+                        type="text"
+                        placeholder="Email"
+                    />
+                    <input
+                        onChange={(e) => setPassword(e.target.value)}
+                        type="password"
+                        placeholder="Password"
+                    />
+                    <button className="bg-green-600 text-white font-bold cursor-pointer px-6 py-2">
+                        Register
+                    </button>
+                    {error && (
+                        <div className="bg-red-500 text-white w-fit text-sm py-1 px-3 rounded-md mt-2">
+                            {error}
+                        </div>
+                    )}
+                    <Link className="text-sm mt-3 text-right" href={"/login"}>
+                        Already registered?
+                        <span className="underline">Log in here</span>
+                    </Link>
+                </form>
+            </div>
+        </div>
+    );
+};
+
+export default RegisterForm;
