+++ app/login/page.tsx
@@ -1,70 +1,17 @@
-"use client";
-import Link from "next/link";
-import React, { useState } from "react";
-import { signIn } from "next-auth/react";
-import { useRouter } from "next/navigation";
+import LoginForm from "../components/LoginForm";
+import { getServerSession } from "next-auth";
+import { redirect } from "next/navigation";
+import { authOptions } from "../api/auth/[...nextauth]/route";
+import React from "react";
 
-const Login = () => {
-    const [email, setEmail] = useState("");
-    const [password, setPassword] = useState("");
-    const [error, setError] = useState("");
+export default async function Home() {
+    const session = await getServerSession(authOptions);
 
-    const router = useRouter();
-
-    const handleSubmit = async (e) => {
-        e.preventDefault();
-
-        try {
-            const res = await signIn("credentials", {
-                email,
-                password,
-                redirect: false,
-            });
-
-            if (res.error) {
-                setError("Invalid Credentials");
-                return;
-            }
-            router.replace("profile");
-        } catch (error) {
-            console.log(error);
-        }
-    };
+    if (session) redirect("/profile");
 
     return (
-        <div className="grid place-items-center h-screen">
-            <div className="shadow-lg p-5 rounded-lgshadow-lg p-5 rounded-lg border-t-4 border-green-400">
-                <h1 className="text-xl font-bold my-4">Log in:</h1>
-                <form onSubmit={handleSubmit} className="flex flex-col gap-3">
-                    <input
-                        onChange={(e) => setEmail(e.target.value)}
-                        type="text"
-                        placeholder="Email"
-                    />
-                    <input
-                        onChange={(e) => setPassword(e.target.value)}
-                        type="password"
-                        placeholder="Password"
-                    />
-                    <button className="bg-green-600 text-white font-bold cursor-pointer px-6 py-2">
-                        Login
-                    </button>
-                    {error && (
-                        <div className="bg-red-500 text-white w-fit text-sm py-1 px-3 rounded-md mt-2">
-                            {error}
-                        </div>
-                    )}
-                    <Link
-                        className="text-sm mt-3 text-right"
-                        href={"/register"}
-                    >
-                        No account yet?{" "}
-                        <span className="underline">Register here</span>
-                    </Link>
-                </form>
-            </div>
-        </div>
+        <main>
+            <LoginForm />
+        </main>
     );
-};
-
-export default Login;
+}
